SDK import
Attempted and failed to integrate the UniMag SDK in Android. The .Jar is
included in the libs folder, and there are a skeleton .h/.m and .java
files in src. I wish future semesters luck in implemeting the swipe
functionality, as it's all that's left with regards to React Native base
functionality.